Update Core version to use abstract class for DI
