+++ src/Microsoft.AspNet.OData.Shared/Extensions/ContainerBuilderExtensions.cs
@@ -52,7 +52,7 @@ namespace Microsoft.AspNet.OData.Extensions
             builder.AddService<TopQueryValidator>(ServiceLifetime.Singleton);
 
             // QueryOption
-            builder.AddService<DefaultSkipTokenImplementation>(ServiceLifetime.Singleton);
+            builder.AddService<SkipTokenHandler, DefaultSkipTokenHandler>(ServiceLifetime.Singleton);
 
             // SerializerProvider and DeserializerProvider.
             builder.AddService<ODataSerializerProvider, DefaultODataSerializerProvider>(ServiceLifetime.Singleton);
