+++ src/Microsoft.AspNet.OData.Shared/Formatter/ODataOutputFormatterHelper.cs
@@ -185,7 +185,7 @@ namespace Microsoft.AspNet.OData.Formatter
                 writeContext.Path = path;
                 writeContext.MetadataLevel = metadataLevel;
                 writeContext.SelectExpandClause = internalRequest.Context.SelectExpandClause;
-                writeContext.OrderByClause = internalRequest.Context.OrderByClause;
+                writeContext.QueryOptions = internalRequest.Context.QueryOptions;
 
                 serializer.WriteObject(value, type, messageWriter, writeContext);
             }
