+++ src/Microsoft.AspNet.OData.Shared/Microsoft.AspNet.OData.Shared.projitems
@@ -64,11 +64,11 @@
     <Compile Include="$(MSBuildThisFileDirectory)Formatter\QueryStringMediaTypeMapping.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Formatter\Serialization\ODataSerializerProvider.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)GetNextPageHelper.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Interfaces\ISkipHandler.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)IPerRouteContainer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ODataNullValueMessageHandler.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PerRouteContainerBase.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Query\DefaultSkipTokenImplementation.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Query\DefaultSkipTokenHandler.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Query\SkipTokenHandler.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\Expressions\DynamicTypeWrapperConverter.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\Expressions\PropertyContainer.generated.cs">
       <AutoGen>True</AutoGen>
@@ -76,6 +76,7 @@
       <DependentUpon>PropertyContainer.generated.tt</DependentUpon>
     </Compile>
     <Compile Include="$(MSBuildThisFileDirectory)Query\Expressions\SelectExpandWrapperOfT.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Query\Interfaces\ISkipTokenConsumer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\SkipTokenQueryOption.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\Validators\SkipTokenQueryValidator.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)RequestMethod.cs" />
@@ -169,8 +170,8 @@
     <Compile Include="$(MSBuildThisFileDirectory)Query\Expressions\UriFunctionsBinder.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\Expressions\IdentityPropertyMapper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\DefaultQuerySettings.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Query\IPropertyMapper.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Query\ICountOptionCollection.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Query\Interfaces\IPropertyMapper.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Query\Interfaces\ICountOptionCollection.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\ModelBoundQuerySettings.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\NotCountableAttribute.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\NotFilterableAttribute.cs" />
@@ -228,14 +229,14 @@
     <Compile Include="$(MSBuildThisFileDirectory)ODataUntypedActionParameters.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)QueryableRestrictions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)QueryableRestrictionsAnnotation.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Query\ISelectExpandWrapper.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Query\Interfaces\ISelectExpandWrapper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\Expressions\NamedPropertyExpression.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\Expressions\PropertyContainer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\Expressions\SelectExpandBinder.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\Expressions\SelectExpandWrapper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\Expressions\ModelContainer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\Expressions\SelectExpandWrapperConverter.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Query\ITruncatedCollection.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Query\Interfaces\ITruncatedCollection.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\TruncatedCollectionOfT.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\SelectExpandQueryOption.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\Validators\SelectExpandQueryValidator.cs" />
