--- src/Microsoft.AspNet.OData.Shared/Query/DefaultSkipTokenImplementation.cs
@@ -1,178 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation.  All rights reserved.
-// Licensed under the MIT License.  See License.txt in the project root for license information.
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Linq.Expressions;
-using Microsoft.AspNet.OData.Common;
-using Microsoft.AspNet.OData.Formatter;
-using Microsoft.AspNet.OData.Interfaces;
-using Microsoft.AspNet.OData.Query.Expressions;
-using Microsoft.OData;
-using Microsoft.OData.Edm;
-using Microsoft.OData.UriParser;
-
-namespace Microsoft.AspNet.OData.Query
-{
-    /// <summary>
-    /// Default implementation for skip token
-    /// </summary>
-    public class DefaultSkipTokenImplementation : ISkipTokenHandler
-    {
-        private IDictionary<string, object> _propertyValuePairs;
-        private const char CommaDelimiter = ',';
-
-        /// <summary>
-        /// Initialize a new instance of <see cref="DefaultSkipTokenImplementation"/>
-        /// </summary>
-        public DefaultSkipTokenImplementation()
-        {
-            PropertyDelimiter = ':';
-        }
-
-        /// <summary>
-        /// Process SkipToken Value to create 
-        /// </summary>
-        /// <param name="rawValue"></param>
-        public IDictionary<string, object> ProcessSkipTokenValue(string rawValue)
-        {
-            _propertyValuePairs = new Dictionary<string, object>();
-            string[] keyValues = rawValue.Split(CommaDelimiter);
-            foreach (string keyAndValue in keyValues)
-            {
-                string[] pieces = keyAndValue.Split(new char[] { PropertyDelimiter }, 2);
-                if (pieces.Length > 1 && !String.IsNullOrWhiteSpace(pieces[1]))
-                {
-                    object value = null;
-                    if (pieces[1].StartsWith("'enumType'"))
-                    {
-                        string enumValue = pieces[1].Remove(0, 10);
-                        IEdmTypeReference type = EdmLibHelpers.GetTypeReferenceOfProperty(Context.Model, Context.ElementClrType, pieces[0]);
-                        value = ODataUriUtils.ConvertFromUriLiteral(enumValue, ODataVersion.V401, Context.Model, type);
-                    }
-                    else
-                    {
-                        value = ODataUriUtils.ConvertFromUriLiteral(pieces[1], ODataVersion.V401);
-                    }
-                    if (!String.IsNullOrWhiteSpace(pieces[0]))
-                    {
-                        _propertyValuePairs.Add(pieces[0], value);
-                    }
-                }
-            }
-            return _propertyValuePairs;
-        }
-
-        /// <summary>
-        /// Gets the given <see cref="ODataQueryContext"/>.
-        /// </summary>
-        public ODataQueryContext Context { get; set; }
-
-        /// <summary>
-        /// Delimiter used to separate property and value, exposing for the purpose of testing
-        /// </summary>
-        public char PropertyDelimiter { get; set; }
-
-        /// <summary>
-        /// Returns a function that converts an object to a skiptoken value string
-        /// </summary>
-        /// <param name="lastMember"> Object based on which SkipToken value will be generated.</param>
-        /// <param name="model">The edm model.</param>
-        /// <param name="orderByNodes">QueryOption </param>
-        /// <returns></returns>
-        public string GenerateSkipTokenValue(Object lastMember, IEdmModel model, IList<OrderByNode> orderByNodes)
-        {
-            object value;
-            if (lastMember == null)
-            {
-                return String.Empty;
-            }
-            IEnumerable<IEdmProperty> propertiesForSkipToken = GetPropertiesForSkipToken(lastMember, model, orderByNodes);
-
-            String skipTokenvalue = String.Empty;
-            if (propertiesForSkipToken == null)
-            {
-                return skipTokenvalue;
-            }
-
-            int count = 0;
-            int lastIndex = propertiesForSkipToken.Count() - 1;
-            foreach (IEdmProperty property in propertiesForSkipToken)
-            {
-                bool islast = count == lastIndex;
-                IEdmStructuredObject obj = lastMember as IEdmStructuredObject;
-                if (obj != null)
-                {
-                    obj.TryGetPropertyValue(property.Name, out value);
-                }
-                else
-                {
-                    value = lastMember.GetType().GetProperty(property.Name).GetValue(lastMember);
-                }
-
-                String uriLiteral = String.Empty;
-                if (value == null)
-                {
-                    uriLiteral = ODataUriUtils.ConvertToUriLiteral(value, ODataVersion.V401);
-                }
-                else if (TypeHelper.IsEnum(value.GetType()))
-                {
-                    ODataEnumValue enumValue = new ODataEnumValue(value.ToString(), value.GetType().FullName);
-                    uriLiteral = ODataUriUtils.ConvertToUriLiteral(enumValue, ODataVersion.V401, model);
-                    uriLiteral = "'enumType'" + uriLiteral;
-                }
-                else
-                {
-                    uriLiteral = ODataUriUtils.ConvertToUriLiteral(value, ODataVersion.V401, model);
-                }
-                skipTokenvalue += property.Name + PropertyDelimiter + uriLiteral + (islast ? String.Empty : CommaDelimiter.ToString());
-                count++;
-            }
-            return skipTokenvalue;
-        }
-
-        private static IEnumerable<IEdmProperty> GetPropertiesForSkipToken(object lastMember, IEdmModel model, IList<OrderByNode> orderByNodes)
-        {
-            IEdmType edmType = GetTypeFromObject(lastMember, model);
-            IEdmEntityType entity = edmType as IEdmEntityType;
-            if (entity == null)
-            {
-                return null;
-            }
-
-            IList<IEdmProperty> key = entity.Key().AsIList<IEdmProperty>();
-            if (orderByNodes != null)
-            {
-                OrderByOpenPropertyNode orderByOpenType = orderByNodes.OfType<OrderByOpenPropertyNode>().LastOrDefault();
-                if (orderByOpenType != null)
-                {
-                    //SkipToken will not support ordering on dynamic properties
-                    return null;
-                }
-
-                IList<IEdmProperty> orderByProps = orderByNodes.OfType<OrderByPropertyNode>().Select(p => p.Property).AsIList();
-                foreach (IEdmProperty subKey in key)
-                {
-                    orderByProps.Add(subKey);
-                }
-
-                return orderByProps.AsEnumerable();
-            }
-            return key.AsEnumerable();
-        }
-
-        private static IEdmType GetTypeFromObject(object obj, IEdmModel model)
-        {
-            SelectExpandWrapper selectExpand = obj as SelectExpandWrapper;
-            if (selectExpand != null)
-            {
-                IEdmTypeReference typeReference = selectExpand.GetEdmType();
-                return typeReference.Definition;
-            }
-
-            Type clrType = obj.GetType();
-            return EdmLibHelpers.GetEdmType(model, clrType);
-        }
-    }
-}
