+++ src/Microsoft.AspNet.OData.Shared/Query/Interfaces/ICountOptionCollection.cs
