+++ src/Microsoft.AspNet.OData.Shared/Query/Interfaces/ITruncatedCollection.cs
